+++ src/cse-machine/interpreter.ts
@@ -13,7 +13,8 @@ import { UNKNOWN_LOCATION } from '../constants'
 import * as errors from '../errors/errors'
 import { RuntimeSourceError } from '../errors/runtimeSourceError'
 import { checkEditorBreakpoints } from '../stdlib/inspector'
-import type { Context, ContiguousArrayElements, Result, Value, StatementSequence } from '../types'
+import type { Context, Result, Value, StatementSequence } from '../types'
+import type { ContiguousArrayElements } from '../stepper/util'
 import * as ast from '../utils/ast/astCreator'
 import { filterImportDeclarations } from '../utils/ast/helpers'
 import { evaluateBinaryExpression, evaluateUnaryExpression } from '../utils/operators'
