+++ src/stdlib/parser.ts
@@ -3,7 +3,8 @@ import * as es from 'estree'
 import { parse as sourceParse } from '../parser/parser'
 import { SourceParser } from '../parser/source'
 import { libraryParserLanguage } from '../parser/source/syntax'
-import { Context, ContiguousArrayElements, Node, StatementSequence, Value } from '../types'
+import { Context, Node, StatementSequence, Value } from '../types'
+import { ContiguousArrayElements } from '../stepper/util'
 import { oneLine } from '../utils/formatters'
 import { vector_to_list } from './list'
